Move Limit pushdown rule to distributed
Summary:
Previously, limit pushdown was in the non-distributed part of the compiler. However, it needs to move to the presplit optimizer, because it needs to be aware of PEMs vs Kelvin for where the limit is executing. This is to prevent the limit from being pushed earlier than a PEM-only UDF, which would result in the PEM-only UDF being scheduled on a Kelvin node. When PEM-only UDFs are scheduled on Kelvin nodes, they return empty strings rather than the desired data, so it is important that limits don't get pushed earlier than PEM-only UDFs.

We are likely to reorganize our rules in the compiler but that will come in a subsequent diff.

Note that this does not add the rule to a rule executor, so it will not be invoked until that happens in another subsequent diff. In order to safely invoke this rule, we first need to add the annotations to PEM-only UDFs as being PEM-only.

Test Plan: added unit test

Reviewers: philkuz, jamesbartlett

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8328

GitOrigin-RevId: 24930bec8cc9b129c1e98d2365654bc521b60f35