Recycle ConnTrackers
Summary:
Once noticed a lot of new/delete calls when profiling with `perf`. Since ConnTrackers are created/destroyed a lot, they are a good candidate for object recycling, so we don't have to incur the memory allocation costs.

Uses C++ "placement new" to re-initialize the memory without performing a memory allocation.

Performance on dev-cluster-oazizi is about 9% and 6%, for the two nodes. The 9% *might* be a slight improvement, since was seeing 10% before.

Will wait for @jps scripts to make proper measurements and see if this is helping or not.

Test Plan: Existing tests - no functional change.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao, jps

Differential Revision: https://phab.corp.pixielabs.ai/D8145

GitOrigin-RevId: b2fcbe3df8261d0906a3ab8df650c9a779fb65cb