Fix CLI release build stashing
Summary:
in D8354, i updated the release builds to use gcs for stashing. the CLI build was failing because another issue (fixed in D8356), so it didn't get to the stashing part yet.
turns out the gcs stashing doesn't work on the mac worker, so we should continue using the regular stash for the mac worker.

after this diff, the CLI release build is finally actually fixed

Test Plan: ran cli rc build

Reviewers: zasgar, vihang

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8359

GitOrigin-RevId: f5c98ae5c6043732c5c8c54f535c3bedba5dcb23