Combine cloudapipb and "public" cloudapipb into a single version
Summary:
Managing a "public" and internal version of cloudapibp was causing
split brain. Our cloud endpoints should be considered public anyway since folks
might want to interact with pixie cloud in other ways beyond what we expose
from the go/py API.

This gets rid of the "public" version and replaces it with the internal version.
Note that this is a breaking change and will break old consumers of this service.

While we don't plan on making such breaking changes, this choice was a difficult
one, and given the fact that we don't have many consumers relying on the API, we
are using this moment to clean it up before it's too late.

Test Plan: Existing tests.

Reviewers: zasgar, michelle

Reviewed By: zasgar, michelle

JIRA Issues: PP-2637

Differential Revision: https://phab.corp.pixielabs.ai/D8365

GitOrigin-RevId: 3a062737dbc0d26c08457a8775b7a46a8199d1e9