[PC-756] Allow users to login to orgs that they already belong to
Summary:
Tested the full invite flow for Kratos/Hydra and found that users couldn't log in to orgs that they were invited to, their domain name overrode whatever settings we already had.
This fixes that and checks the user's org information first, using that to login if it exists.

Test Plan: Added a test for the functionality, updated existing tests, and tested on skaffold deployed cloud - successfully logged in.

Reviewers: michelle, vihang, zasgar, nserrino

Reviewed By: michelle

JIRA Issues: PC-756

Differential Revision: https://phab.corp.pixielabs.ai/D8363

GitOrigin-RevId: 0576f48a9c9385df97f3bcfe17a531351c01a9c5