PP-2650 Add initial boilerplate for operator
Summary:
Adding initial boilerplate for our Vizier operator, based off of the generated kubebuilder project.
kubebuilder generated:
- contents of operator directory
- k8s/operator/crd
I heavily modified the contents of the operator directory to remove the additional items we didnt need and made it work with our build system. removed most of the items in the Makefile except for the commands that help generate the k8s crd yaml and generated go files.

Test Plan: deployed the crd, deployed the operator, and deployed an instance of the Vizier custom resource. verified that the operator detected the change and triggered the "Reconcile" function.

Reviewers: zasgar, vihang, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

JIRA Issues: PP-2650

Differential Revision: https://phab.corp.pixielabs.ai/D8367

GitOrigin-RevId: 2c0ae0e45c97309258f083eea5d33acb2fc15578