PP-2683 Don't hardcode namespace in Vizier
Summary:
We are updating Vizier so that it can run in any namespace. the first step to doing so is to update vizier so that it does not rely on any hardcoded "pl"s.
instead, we should use the namespace that we can get from the pod metadata.

Test Plan: created an rc, deployed it. updated vizier running on actual release version. verified all cases continue to work/updates worked.

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2683

Differential Revision: https://phab.corp.pixielabs.ai/D8420

GitOrigin-RevId: dc945ad85075052f8322e48fd4d8224332b521af