PC-874 Make vizier-image-secrets optional
Summary:
Currently the vizier-image-secrets are used in two places:
1. In cli deploys, when deploying vizier versions older than 0.5.22
2. In our template generator, which is run during releases.

the vizier-image-secrets are used in the API server's `GetImageCredentials` rpc call which only the CLI deploy calls.
given that we are now on vizier 0.7.2, we can safely clean up the CLI's handling for viziers older than 0.5.22. since this is the only consumer, we can likely also remove the `GetImageCredentials` call itself. for now I've just made the vizier-image-secrets optional, and that call will error if they dont exist.

Test Plan: ran `px deploy`

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8486

GitOrigin-RevId: a2144f54d08a444334b54bcabda65d16365a8a3d