ensure_namespace now doesn't exit if set -e is called
Summary: set -e with ensure_namespace would exit unexpectedly if ns was missing. kubectl get ns returns an error code. Now we capture it and return foo in case of exit code. Migth need to clean up, dependinag on reviewers opinion.

Test Plan: tested to make sure that if ns exists, it doesn't try to create one and if it doesn't, under set -e we don't exit the containing script. The script i used to test was the cloud_prereqs one.

Reviewers: vihang, michelle

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8483

GitOrigin-RevId: 8c8eb66f4b3ce1ea2ecfe454164a1823b7d25e05