PC-870 Don't require cloud-dns-secrets in dnsmgr
Summary:
we currently require users to enter cloud-dns-secrets to add DNS records for users using passthrough.
this diff updates the dnsmgr deployment so that supplying these secrets is optional. in this case, direct mode won't work properly. however, the assumption is that if they are hosting their own cloud, this is less of an issue.

there is still some inefficiency with how vzmgr requests the DNSAddress from dnsmgr. once the vizier's proxy server is ready with an IP address, it will make a request to dnsmgr every heartbeat.
i thought through a couple of ways to clean this up, for example, only making this call only when the user has direct mode enabled. but we probably need something a little more complex to satisfy these conditions:
1. DNS records take some time to propagate. if we only make the request once the user enables direct mode, there will be some period of time where they can't query their cluster.
2. we need to handle cases where the user's proxy IP address has changed, which is why we continue to request a DNS address even after the vizier already has an addressed assigned. it is currently dnsmgr's responsibility to handle changing IP addresses
given this, we probably need to think more carefully about the fix for this. for the purposes of this issue, we should just make the secrets optional.

Test Plan: test with staging

Reviewers: vihang, zasgar

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8492

GitOrigin-RevId: cf87c5349f0e2cd736c1c8ae0cb286530ec1fd1e