PP-2713: Move ORY migrate.yaml code in to deployment for hydra/kratos
Summary: Migrate code was in load_dev_db.sh, move to initContainers for deployment. The mgirate code is idempotent so not a problem if it runs many times.

Test Plan: Tested on a fresh cluster and a running cluster. Hydra/Kratos migrations still occur, can confirm with psql on the postgres pod.

Reviewers: vihang, michelle

Reviewed By: vihang

JIRA Issues: PP-2713

Differential Revision: https://phab.corp.pixielabs.ai/D8508

GitOrigin-RevId: 4a71b329064d9f30a446ea1130f94c08d87780fb