[PC-674] Always use `pl-history`, instead of sidestepping it.
Summary: Doing this required patching a few things to match it. Additionally, it revealed an existing synchronization issue between `ScriptLoader` and `ScriptContext`, which this diff also fixes. There should be fewer instances of getting blank table/chart boxes after a script runs because of that.

Test Plan: Poke around the Live view. Try going to `/live` directly, refresh the page with a few scripts, run them again without a refresh. Analytics from `pl-history` should pick up on all of the URL transitions that it didn't before. The script results shouldn't resolve as blank so frequently (there seem to be other ways for this to happen, not addressed here).

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-674

Differential Revision: https://phab.corp.pixielabs.ai/D8537

GitOrigin-RevId: 0d020c7a343890dd505a4124d8264e8757f95275