PP-2729 Update cert-provisioner to check jwt-signing-key
Summary:
we have a current deploy/delete flow which is broken.
1. user deploys helm chart
2. user deletes helm chart. this leaves any generated certs, because they are not tagged with the helm annotations that helm uses to figure out what it should delete.
3. user deploys helm chart again.
4. cert-provisioner job sees that the proxy-tls-certs exist. doesnt update the secrets as it should.

one solution for this would be to update the generated secrets so that they have the correct helm annotation. that would require some extra work, such as detecting the user has deployed through helm and what that exact annotation should be.
since this bug will be solved by introducing the operator, I figured it wouldn't be worth that work. instead this simple fix will solve the problem in the meantime.

Test Plan: created rc. generated yamls with --extract_yaml command. deployed the YAMLs, then `kubectl delete -f` the YAMLs. confirm that it leaves the `proxy-tls-cert` in the namespace. redeploy by running `kubectl apply` and confirm that cert-provisioner job generates new secrets. also update existing deployment to the new rc, verifies it does not generate new secrets in this case.

Reviewers: nserrino, vihang

Reviewed By: nserrino, vihang

JIRA Issues: PP-2729

Differential Revision: https://phab.corp.pixielabs.ai/D8546

GitOrigin-RevId: b82c5b8a4bdc1965691fe70e528429400c92b636