ConnStats 2.0: New BPF code
Summary:
This is the new BPF code for reporting ConnStats. Note that the reported events are not hooked up yet in user-space, so while this code is active, it goes into a black-hole.

This is done to stage out the diff.

Test Plan: Existing tests to make sure nothing is broken. In future diffs--once things are hooked up--conn_stats_bpf_test will ensure this is correct.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D8548

GitOrigin-RevId: a227b8d2465db141b99f8a78c0d035a0fb02a20f