PP-2736 Update vzmgr to read disableAutoUpdate from heartbeat
Summary:
we are moving the auto update config to vizier. vizier communicates the new autoupdate setting in its heartbeats.
this diff updates vzmgr so that it uses the autoupdates from the heartbeat.

Test Plan: deployed to staging and built an rc. verified that updating the rc configmap for disableAutoUpdate works

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2736

Differential Revision: https://phab.corp.pixielabs.ai/D8606

GitOrigin-RevId: f36922f04192d31cd5985e07226dad0aef7a64b9