Don't set NATS pendingLimits on Chan subscriptions
Summary:
Setting pendingLimits on Chan subs is disallowed.
https://github.com/nats-io/nats.go/blame/6346f3830b6606227dfb8c6dddcb62d204d39fa0/nats.go#L4134-L4136
This was added in D3739 but the returned error was silently ignored.

We started checking and logging the returned error (if any) in D7960 and now
this error shows up in logs and ends up being a red herring when trying to debug
other issues.

Test Plan: N/A

Reviewers: zasgar, michelle

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D8604

GitOrigin-RevId: dea749ba93cab844486a9c094148e11999a04715