Get rid of storybook
Summary:
storybook is a great idea in theory and would be useful if we had
many ui components. However it is currently not used and ends up causing us
to download a lot of npm deps and having to constantly address CVEs in those deps.

Let's remove storybook for now. We can consider reviving this in the future
if we see a good usecase for it.

Test Plan: N/A

Reviewers: #third_party_approvers, zasgar

Reviewed By: #third_party_approvers, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D8614

GitOrigin-RevId: 104b3018190d78595fdac5117b9cd7bd1289a937