PC-895: Auto-approved invited users.
Summary:
Users who are invited should be auto-approved when they're invited.

Another option is to add a nullable bool field to CreateUser so that this is a bit cleaner, but it does muddy the CreateUser experience.

Test Plan: Undisabled test and updated the test with encessary fields, then re-disabled to prevent flakes.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-895

Differential Revision: https://phab.corp.pixielabs.ai/D8621

GitOrigin-RevId: fe2feb054a327ff1cff9afb1ddacb27309d814c5