[PC-917] Start overhaul of URL state management (feature flagged)
Summary:
WARNING: In this diff and those based on it, everything is locked behind the `USE_SIMPLIFIED_ROUTING` feature flag in `vizier.tsx`. Do not submit with that flag set to true; do set it to true for testing though.

Implements `VizierRoutingContext`, a context that handles the URL and parameters therein such that no other code should need to interact with the URL directly. Begins to wire updated versions of `ScriptContext` and so forth to use this simplified model. So far, it only handles the state management relating to the active script and the URL. Implementation for running the script (`ScriptLoader` and `ScriptContext.execute`) is just skeletal so far; numerous contexts still need to be ported.

For any updated components/contexts/etc, please create a `new-whatever-context.tsx` beside the original that contains the changes. Use the new version in the `USE_SIMPLIFIED_ROUTING` path. A final diff once everything works again will remove the flag, call the new path only, delete the old versions of everything, and rename the new versions to take their place.

See [PC-917](https://pixie-labs.atlassian.net/browse/PC-917) for details of what still needs to be done. TODOs are sprinkled generously throughout this diff as well.

Test Plan:
First, ensure behavior is unchanged when `USE_SIMPLIFIED_ROUTING = false`.

Then set it to `true`, and visit `withpixie.ai:8080/`. Watch as the URL and the console. The URL should update automatically to choose a default cluster in the same manner we do now; script memory between sessions should be gone; the `Try Another` button should change the URL; the console should emit a debug message each time the script/args/cluster change - but only once per combination of changes. These methods will be the basis for all new script-related state+url management.

Reviewers: michelle

Reviewed By: michelle

Subscribers: vihang, philkuz

JIRA Issues: PC-917

Differential Revision: https://phab.corp.pixielabs.ai/D8629

GitOrigin-RevId: 820fb6746f443aae4fe27571e1c6a05b6d811044