UDF for HTTP Response message
Summary:
HTTP table's resp_message column is going to go away. The number of users of the column are likely very low, and it's redundant information.

As a courtesy, we'll provide a UDF to convert the status codes to messages (e.g. 404 -> Not Found).

I followed the pattern of `net` for this, but it might be better as part of builtins. Please let me know the preferred location and naming for consistency.

Test Plan: Test added

Reviewers: philkuz, nserrino

Reviewed By: philkuz

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8630

GitOrigin-RevId: 835cbfe3f4f4f31fe1afd93cf3c63e5be10e0b1f