PC-928 Fix lint issues in UI
Summary:
These are lint issues that are triggered by the changes in D8632. It's better to fix them separately to keep that diff size as small as possible.
there are no logic changes here, just movement around of code because of complaints where a component was used before it was defined, addition of default args, etc.

Test Plan: yarn dev

Reviewers: nlanam, vihang, philkuz, zasgar

Reviewed By: vihang

JIRA Issues: PC-928

Differential Revision: https://phab.corp.pixielabs.ai/D8637

GitOrigin-RevId: affa69333ee5300dfd4f6bc525da40f45bcf34a8