PC-937 - Add esm transformation for jest
Summary:
This is necessary because we are upgradeing to mui@5 which uses
esm modules. ESM in node modules are not transformed by default.
Unfortunately, changing the tranform rules is not enough since
some Babel config is preventing the transformation without the
extra level of indirection.

Test Plan: tested with current version and update mui.

Reviewers: michelle, vihang, nlanam

Reviewed By: vihang

JIRA Issues: PC-937

Differential Revision: https://phab.corp.pixielabs.ai/D8651

GitOrigin-RevId: a6d0a2b497b88e02222ccc771867008fed108c56