Updates to stan subscriptions for indexer
Summary:
This is a durable queue subscription so the subscription should resume where we left off.
The only time we want `DeliverAll` is when we update the index version and would like to reindex everything available.
Using a queue based on the index version helps achieve this behavior.

Also add a MaxInflight to add some backpressure to stan message delivery.

Test Plan: dev cloud

Reviewers: zasgar, michelle

Reviewed By: michelle

JIRA Issues: PP-2743

Differential Revision: https://phab.corp.pixielabs.ai/D8658

GitOrigin-RevId: a6dabb3a18257d4f1066512b771c92571473820f