PC-881,PC-942: Add cluster loading component, rely on better logic
Summary: Cluster loading component was not acting consistenly with the old stuff, so I fixed it and probably fixed PC-881. Basically switched over from some janky timer component to instead check to see whether a user has run a script on the cluster during the current LiveView session. If they have not, we will show a loading circle. If the cluster is unhealthy, we show an Alert.

Test Plan:
tested on yarn dev w/ 3 types of clusters
1. Healthy cluster, from start.
  > shows a ClusterConnecting dialog, then goes away after data loads.
2. Healthy cluster, edit px/namespace then rerun
  > Does not show clusterconnecting dialog, should just change layout
3. Disconnected cluster
  > shows an unhealthy alert, never the diagram.
4. Switch from healthy cluster to healthy cluster
  > same as 1

Reviewers: nlanam, vihang, michelle

Reviewed By: vihang

JIRA Issues: PC-942, PC-881

Differential Revision: https://phab.corp.pixielabs.ai/D8663

GitOrigin-RevId: 355c2e1f4ca51936c85b6528a86f67e4436d8e2f