Add Stringifier & its test to fix 'empty stack traces' in the perf profiler.
Summary: In the stirling perf profiler, add a Stringifier class to independently memoize folded symbolic stack traces for user & kernel stack-ids. Fixes the bug of 'empty stack traces' that came as a result of removing stack-ids from the stack table as we iterated through histogram (and then later finding, e.g., the same user stack-id paired with a different kernel stack-id but not being able to read the stack trace addrs because the user stack-id was already removed).

Test Plan: stringifier_test (added in this patch) and existing tests.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2688

Differential Revision: https://phab.corp.pixielabs.ai/D8635

GitOrigin-RevId: 4cf38512bdb69bee15966173c71cfe2936030353