Update most cases of useListCluster to only fetch necessary fields
Summary:
we only need all fields in useListCluster in the admin page... the rest of the pages actually only care about id/name/status.
this diff updates useListCluster so that you can query for the full set of data, or just the necessary subset.
after D8691, more of these can change to a single getClusterInfoByName call.

Test Plan: yarn dev

Reviewers: vihang, philkuz

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D8693

GitOrigin-RevId: 432d4f1df10a703cf8a6bebed329f7a705765c47