[PC-953] Add footer to all pages that didn't already have it
Summary: Auth had it already. Credits, admin, and live did not. Made it stick to the bottom of the page if content would not have pushed it further. If content scrolls, footer ends up below it so as not to waste screen real estate. Also renames the component for the footer since it's general-use now.

Test Plan: Check live, credits, and admin. In all cases, scrolling to the bottom should show the footer (ts&cs, privacy policy, copyright). If the page is shorter than the viewport, the footer should push itself to the bottom of the viewport (credits page does this, as does the scratch pad script before putting code into it).

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-953

Differential Revision: https://phab.corp.pixielabs.ai/D8722

GitOrigin-RevId: 66d73f826ae98547e8b0aff8a6ff6bd1db05cced