[PC-956] Restore live tour functionality
Summary:
Fixes a sizing bug in the menu, restores the context needed for the live tour to open, fixes a logic flaw in `toggleSidebar` for the same purpose, then moves punchouts and tips around in the tour to mostly suit the new UI layout. Not perfect, but good enough for a demo tomorrow.

{F157645}

Test Plan: Open the live tour. You may need to wait a few seconds before the button is responsive (seems to be an unrelated bug).

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-956

Differential Revision: https://phab.corp.pixielabs.ai/D8726

GitOrigin-RevId: acd5df6528a5bfcd3bfd1032e21aeaa2ba36b10a