PC-961: Render an error when a required argument is missing
Summary:
We had an unnecessary check in our script-context to make sure args are valid. We do this check in execute() anyways and so remove the one in script-context.
Also updated the error rendering to be in resultsContext rather than a bar.
{F157703}

Test Plan: Tested on yarn dev and now px/pod can be navigated too without a silent error.

Reviewers: vihang, nlanam, michelle

Reviewed By: vihang, michelle

JIRA Issues: PC-961

Differential Revision: https://phab.corp.pixielabs.ai/D8736

GitOrigin-RevId: e0e9e6e435a338710116d110d0f38170011ca961