PC-967: Update admin/cluster-details GraphQL queries  to explicitly invoke useQuery
Summary:
Explicitly call GraphQL queries to save some energy on what data we grab and push graphQL controls up to the developer rather than hiding behind the interface.
Also took the liberty of cleaning up the cluster-details code to remove funky re-renders. Basically, we should re-use the navigation components and not have to refresh the page on each click by enabling better composure.

Also ended up fixing a bug where clusterNames did not show up in the breadcrumbs. This is because the underlying gql query changed, and the query was the wrong one in the first place.

Test Plan: Tested on yarn dev

Reviewers: michelle, vihang, nlanam

Reviewed By: michelle, vihang

JIRA Issues: PC-967

Differential Revision: https://phab.corp.pixielabs.ai/D8770

GitOrigin-RevId: 8559debeb82e6b829b0bcc207f38c2683a746d2a