PC-970 Fix script execution for OSS cloud
Summary:
With the recent script-context refactor we also switched script execution over to our new API Client, which doesn't have any handling for dev cloud.
for dev cloud, we need to point the passthrough URL to :4444 (envoy) since there is no loadbalancer to do that for us.
we have decided to move a simplified version of pixie-api back into the main package. once that work is underway, it'll be easier to incorporate this change into there.

Test Plan: yarn dev

Reviewers: nlanam, vihang, philkuz

Reviewed By: vihang

JIRA Issues: PC-970

Differential Revision: https://phab.corp.pixielabs.ai/D8778

GitOrigin-RevId: 7ed59bb2a036188817352bcd22b1f330addca4d1