Fix up live view gql queries and loading statuses
Summary:
This diff started as a way to inline the GQL query in live.tsx
Along the way I noticed a bunch of problems with our loading status
indicator. This fixes all known problems.

Test Plan:
yarn dev
On a disconnected cluster, we suggest the user pick a new cluster.
On an unhealthy cluster, we suggest the user contact us.
On a healthy cluster, we show a loading indicator while we wait for the GRPC
connection to start.
On a healthy cluster with a bad script, we tell the user to pick a new script.
On a healthy cluster with a bad script arg, we render the canvas with the
execution error.
On a healthy cluster with a good script and good args, we render results!

Reviewers: philkuz, zasgar, michelle, nlanam

Reviewed By: philkuz

JIRA Issues: PC-971, PC-966

Differential Revision: https://phab.corp.pixielabs.ai/D8782

GitOrigin-RevId: 48700920b568223ad7009a074575cb0094ef3630