PP-2445: Fix some bugs in distributed analyzer when propagating column types
Summary:
When finishing up PP-2445, some bugs got exposed where the column type was not getting properly propagated. This oversight is in part due to the fact that we have two ways of tracking types right now on tables (relation vs tabletype) which we should consolidate into just tabletype.

Additionally, previously added was a rule to split PEM-only UDFs out from the operator that they are a part of. This is amended in this diff to only do so when the PEM-only UDF is co-located with a Kelvin-only UDF. Splitting the two is necessary in this case, because Kelvin and PEM-only UDFs can't be scheduled on the same operator. But when no Kelvin-only UDF was present, we were being overeager in splitting the operator before this change.

Test Plan: updated tests

Reviewers: philkuz, jamesbartlett, zasgar

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8804

GitOrigin-RevId: ae23485a3089d3abe9ebd8b3ce97d29d54538db9