PP-2445: Hook in relevant rules for distributed analyzer and splitter
Summary: Depends on D8804. This diff hooks in the rules for the presplit analyzer and presplit optimizer. These were added to the distributed splitter in order to support awareness in the planner of UDFs that can only run on PEM or Kelvin, but will be a dedicated place going forward for any analyzer or optimizer step that needs to occur in the distributed phase but before the distributed plan is created.  For the analyzer, this change hooks in the rule to split PEM-only and Kelvin-only UDFs. For the optimizer, it hooks in the rule to push limits earlier in the plan, until a PEM-only UDF is reached.

Test Plan: Added unit tests for both the presplit analyzer and optimizer

Reviewers: philkuz, jamesbartlett, zasgar

Reviewed By: philkuz

JIRA Issues: PP-2445

Differential Revision: https://phab.corp.pixielabs.ai/D8805

GitOrigin-RevId: 81b5256d1d7ec46d7e1b670f9c373032a1fe8953