Move analyzer into its own directory to set the stage for rule re-organization
Summary: As part of PP-2445, we are cleaning up the organization of our rules because they are getting a bit unwieldy in a single file. The plan is to have analyzer-specific rules in the analyzer directory (added by this diff), and optimizer-specific rules in the optimizer directory (already the case). Next diff will be to move the analyzer contents of rules.h into the analyzer directory, with each rule as its own file. The distributed rules will be addressed afterwards.

Test Plan: existing

Reviewers: philkuz, jamesbartlett, zasgar

Reviewed By: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D8807

GitOrigin-RevId: 92a6ff13fb22e709090b2b97d7119810842d1132