Split compiler rules into individual files, grouped by where they are used.
Summary:
As part of PP-2445, we are reorganizing our rules because they have gotten very unwieldy being in one file. The pattern going forward is to split each rule into its own file, and it will live near the component that uses it (rather than a separate rules.h file). This will make the compiler rules more in line with the organization of the distributed and optimizer rules. In the previous configuration, having all of the compiler rules in one place (other than optimizer rules) was confusing because it was unclear where they were used and the files/tests were getting really long.

This is a very large diff but it's just moving existing code around.

Test Plan: existing tests

Reviewers: philkuz, jamesbartlett, zasgar

Reviewed By: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D8814

GitOrigin-RevId: 7e7b74553bc6aa88606fa7c953805771e6aa3b57