PP-2653 Operator: implement updateResourceConfiguration
Summary:
before deploying vizier to a user's cluster, we need to update the configuration to add the user's custom labels, annotations, and resource requirements.
this diff adds the implementation for the functions which add the configurations to the k8s resources.
unfortunately, its not the prettiest since everything is an unstructured object consisting of maps of interfaces. most of the code is checking the map and casting. any suggestions for making this look nicer are appreciated...

`addKeyValueMapToResource` is an exact copy of code that we had before we had templates.

Test Plan: deployed CRD with operator running. confirm that the deployed vizier pods had the annotations and resource requirements

Reviewers: vihang, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2653

Differential Revision: https://phab.corp.pixielabs.ai/D8823

GitOrigin-RevId: eca0c501ace2e2b3c6d8225d148247668947a579