PP-2653 Operator: vizier deploy should update status
Summary:
we should update the vizier status from pending -> running once the deploy is complete.
ideally, we would update it to running once the vizier becomes healthy. however, setting up the healthcheck between the operator and querybroker or cloudconn can be messy because of TLS/auth.
For now, we can just set the status to "running" once the cluster ID is populated and the vizier has successfully connected to the cloud.

Test Plan: ran operator, deployed a vizier. verify the status is set to "Running"

Reviewers: vihang, vjain, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2653

Differential Revision: https://phab.corp.pixielabs.ai/D8842

GitOrigin-RevId: bc0da371da02193eae1eaedd48f666e2482723ec