PP-2793 Move cert-provisioner to operator
Summary:
we used to rely on the cert-provisioner job to generate certs for a vizier when it starts up for the first time.
we can now do this in the operator. this code is mostly a copy from the current cert-provisioner job. once the operator migration is done, we can finally remove the cert provisioner job. for now, it will still start up, but be a no-op.

Test Plan: run operator, deploy new vizier CRD. confirm that vizier starts up properly

Reviewers: vihang, vjain, zasgar, nserrino

Reviewed By: vihang

JIRA Issues: PP-2793

Differential Revision: https://phab.corp.pixielabs.ai/D8870

GitOrigin-RevId: 0201df55b6db1f0bafac9d872a838dd66f25d91a