PP-2654 Add operator handling in artifact tracker
Summary:
we are working on creating a release build for the operator, then allowing the CLI to pull operator artifact versions from Pixie Cloud.
small updates need to be made to the artifact tracker service so that this operation can be done in the OSS cloud.

Test Plan: existing tests pass

Reviewers: vihang, zasgar, vjain

Reviewed By: vihang

JIRA Issues: PP-2654

Differential Revision: https://phab.corp.pixielabs.ai/D8876

GitOrigin-RevId: c4259ea46ef3129dd51527a7a354e79577540ed3