PC-998: Return an Auth0 Identity Provider
Summary: Auth0 sends the idp concatenated as part of the id_token `sub` field. We get the `id_token` from the auth0 /userinfo endpoint. For now this will parse that sub and return the idprovider if it exists in the change.

Test Plan: Added a test to make sure the field is parsed out as long as auth0 returns what we expect.

Reviewers: michelle, zasgar, vjain

Reviewed By: michelle

JIRA Issues: PC-998

Differential Revision: https://phab.corp.pixielabs.ai/D8891

GitOrigin-RevId: e6a28ba08c43ed7f375659c2ae93d0e6cda37641