PC-995: Get Auth0 Identity Provider from a specfic field rather than trying to parse Sub field
Summary: context: I tried to e2e test the Identity Provider code and found that the field I used for the provider info only applied to the id_token, but there happened to be another field called "identities" that contained the provider info. Adds a fix to use that info instead of parsing the "Sub".

Test Plan: Tested on skaffold and updated tests.

Reviewers: michelle, nserrino, vjain, vihang

Reviewed By: michelle

JIRA Issues: PC-995

Differential Revision: https://phab.corp.pixielabs.ai/D8905

GitOrigin-RevId: 4ece13543c3cca85ef09b1f943bdb783809a131b