PP-2798: Quick Fix Limit Pushdown from erroring out around a Union.
Summary:
Limit pushed down around the Union kept failing. Was because we fail on max, but the rule actually continues running. This does raise the question why this happened, but having a fix for now should keep things rolling, especially because this is an optimization.

Minor amounts of investigation shows that limit pushdown will iterate forever on a Union.

Test Plan: Added a test that fails when we want it to fail.

Reviewers: nserrino, vjain, jamesbartlett

Reviewed By: jamesbartlett

JIRA Issues: PP-2798

Differential Revision: https://phab.corp.pixielabs.ai/D8901

GitOrigin-RevId: 16d55f05b6df96119608def2fd652a6fafc136a0