Update cache for settings upon mutation
Summary: It was possible for Apollo to cache the empty string for `tourSeen`, and then never bother to check upstream if that had changed. It then told upstream to save `true`. Telling Apollo how to update its cache as a result of that mutation fixes this without an extra network call.

Test Plan: Try an incognito window on a new account. The tour should only appear once.

Reviewers: michelle, nserrino

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D8928

GitOrigin-RevId: d391ad8d1c15cca5e84cb606fda3ff4e724df3c6