PP-2652 Update Cloudconn to trigger updates on CRD
Summary:
Currently, when cloudconn receives a message from Pixie Cloud to update vizier, cloudconn runs the update by deploying an update job.
Now that we have the operator, we need to tell the operator to update vizier. we do this by updating the Vizier CRD to the new version. the operator is already listening to changes on the CRD, so it begins the update process as soon as the version is changed.

Test Plan:
deploy vizier without operator, but with CRD deployed. trigger an update and confirm that the CRD's version is updated.
with vizier + opreator deployed, triggered an update. confirm the CRD's version is updated and that the operator also updates the vizier to the new version.

Reviewers: vjain, nserrino, zasgar, vihang

Reviewed By: nserrino

JIRA Issues: PP-2652

Differential Revision: https://phab.corp.pixielabs.ai/D8934

GitOrigin-RevId: 7f2d380b031734b66f0798c1bad7ebd8a8428354