PP-2800 Generate OLM operator bundle in operator release build
Summary:
we are using OLM to manage our operators. to do so, we need to generate a bundle in the expected format for OLM.
this diff adds the operator release build, which builds the operator image, then generates a OLM bundle for that image.

Test Plan: deploy OLM to a cluster, verify that the operator build is deployed and is updated when a new bundle is added to the index

Reviewers: zasgar, vihang

Reviewed By: vihang

JIRA Issues: PP-2800

Differential Revision: https://phab.corp.pixielabs.ai/D8932

GitOrigin-RevId: df0527d926e0db40ba612381a14e66d34f62c36c