PP-2652 Cloudconn should still send UpdateResponse for operator-updated vizier
Summary:
previous diff updated cloudconn so that it triggers an update in the operator by updating the CRD version.
however, that diff failed to send a response back to the cloud to let it know the update was successfully received.
if you initiated the update through the cli `px update vizier`, it would just hang since it never receives the response message.

Test Plan: created rc and ran `px update vizier`

Reviewers: vihang, nserrino, vjain, zasgar

Reviewed By: vihang

JIRA Issues: PP-2652

Differential Revision: https://phab.corp.pixielabs.ai/D8951

GitOrigin-RevId: fa2af363bf539859a5879066e9983496dfb97e4d