Make vis follow the spec
Summary:
defaultValue is of type `google.protobuf.StringValue` instead of type
`string` so that there's a distinction between null and empty string.
However the vis.proto wasn't doing the same.

This brings them inline with each other to ease the transition to parsing the
visspec on the UI side.

Test Plan: Ran UI with `yarn dev` and custom script bundle

Reviewers: zasgar, michelle, philkuz, nlanam

Reviewed By: zasgar, nlanam

JIRA Issues: PC-1035

Differential Revision: https://phab.corp.pixielabs.ai/D8965

GitOrigin-RevId: cca6af96ebc3d108ab6bb559661390f7f5950b77