PC-1037: Migration to add auth_provider_id to users table
Summary: add auth_provider_id to the users table as part of our addition of a better id. Used the default value to prevent issues with NULL values in the column, especially because we won't be using this column for a few diffs. ( This happened before when we added `identity_provider` column to the table. We had to fix it with this: https://phab.corp.pixielabs.ai/D8909)

Test Plan: Tested on skaffold to see if it's added. Postgres adds it.

Reviewers: michelle, nserrino, vihang

Reviewed By: vihang

JIRA Issues: PC-1037

Differential Revision: https://phab.corp.pixielabs.ai/D8970

GitOrigin-RevId: 9a7bdf08523db7c1f49974381c6a07fdd746ad62