Always hold onto parsed vis in scripts
Summary:
Holding onto both the parsed and string versions was getting a bit
squirelly and we want the parsed version at almost all times anyway. The only
place that consumes the raw string is the editor and manually stringifying and
parsing at that boundary is easier.

This also eases the transision to proto parsing the Vis (instead of just JSON
parsing).

Test Plan: `yarn dev`

Reviewers: nlanam, philkuz, michelle

Reviewed By: nlanam

Differential Revision: https://phab.corp.pixielabs.ai/D8983

GitOrigin-RevId: abce0b2d8bdb66b8c5c55901529ae378f914363e