[CLEANUP] DataTable set ID
Summary:
This prepares to add SourceConnector::PushData() which move pushing data into SourceConnector.

Since PushData() only takes a vector of DataTable, it still needs a global ID to identify the right
destination. The ID is owned by InfoClassManager, such that when DataTable objects are replaced, the
ID stays the same.

Test Plan: Jenkins

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D8989

GitOrigin-RevId: a9b2d3f8d6104da8cdf8d12ed62d4df29b1324ac