Canonicalize 'start_time' (rather than 'start') in PxL scripts
Summary: In the UI, we have some special handling to propagate the "start_time" parameter of scripts when links are clicked and views are switched. We were inconsistent with the naming in some of our scripts, calling the field "start" in some places. This diff standardizes our scripts on "start_time". Additionally, in the process of updating all of the usage of "start", discovered some stale scripts and deleted them.

Test Plan: ran every view that has changed to ensure compilation and successful execution

Reviewers: vihang, zasgar, michelle

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9031

GitOrigin-RevId: 9e0fe6c7f1922c17fbe51b954905d7e2daaa00a8