PP-2835 Fix request graph hierarchy toggle zoom
Summary:
When toggling between hierarchy mode in the request graph, it doesnt zoom out to capture all of the nodes.
also found a bug where toggling the error/latency views wouldn't save the hierarchy options and just switch the user back to clustered mode.
this updates the hierarchy toggle so that we actually update the graphOpts.

Test Plan:
hierarchy mode:
{F160849}

Reviewers: nlanam, vihang, nserrino

Reviewed By: vihang

JIRA Issues: PP-2835

Differential Revision: https://phab.corp.pixielabs.ai/D9055

GitOrigin-RevId: caa68c9077e4200437a032acd24653513e1b4c47