Make value in breadcrumbs clickable
Summary:
Currently, you have to click the down arrow in the breadcrumbs to open the menu/editor for the breadcrumb.
however, the convention usually is that you can click the breadcrumb value to open up the value editor too (see gcp project selector). this will make the click target for breadcrumbs larger, which is good since the arrow is pretty small

Test Plan: yarn dev

Reviewers: vihang, nlanam

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9056

GitOrigin-RevId: e40a9d54d70f727c6e845d941b439aeca2051ebe