PC-1062 Quickly changing scripts won't cancel execution
Summary:
we were seeing a bug where if you switch scripts or clusters too quickly while the previous script was still executing, the current execution will be cancelled.
this is because we call cancelExecution(), which updates the numExecutionTries to 0, after we've already updated numExecutionTries to 1 (or more, for mutations).
instead, we can fix the order of the calls. we can cancelExecution, then update numExecutionTries after. This allows the following script to actually execute

Test Plan: yarn dev, switch between scripts quickly before they are done executing.

Reviewers: vihang, nlanam, philkuz

Reviewed By: vihang

JIRA Issues: PC-1062

Differential Revision: https://phab.corp.pixielabs.ai/D9063

GitOrigin-RevId: b8c9b042f1199bc010bf21cf438c18466fa119ac