cgroup_path_templater: A new approach to cgroup paths
Summary:
We've been bitten a number of times by systems that follow different naming conventions for their cgroup paths.
When this happens we fail to monitor anything, because our list of PIDs is blank.

To make our approach more robust, this file automatically discovers the appropriate path template to use.
This is in place of us having to manually put the templates in place.

Test Plan:
Note that this is not yet enabled, it just adds the utility.

The utility itself includes tests on different formats added. I also tested manually on a GKE cluster to ensure no regression.

Reviewers: michelle, zasgar

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D9076

GitOrigin-RevId: 05379c5a39537446793a7f3cc238f74aa7fa1a66