[PC-1061] Allow the Pixienaut box to grow wider if needed
Summary:
The CLI auth flow in particular made it weird for the box to be so thin.

Most dialogs still look the same:
{F160994}

CLI token growing to fit:
{F160995}

An error dialog growing wider to fit its text:
{F160993}

What happens if the screen isn't wide enough:
{F160996}

Test Plan: `withpixie.ai:8080/login?local_mode=true`. If you refresh that view, you'll get an error page that is also affected. The sign up and log in boxes are not any wider as they didn't need to be.

Reviewers: #third_party_approvers, michelle, vihang

Reviewed By: #third_party_approvers, michelle

JIRA Issues: PC-1061

Differential Revision: https://phab.corp.pixielabs.ai/D9072

GitOrigin-RevId: f435da62815d4eea0f4f459e37f1a350b2f3aa56