PC-1072: Implement CreateIdentity and CreateInviteLinkForIdentity for kratos
Summary:
Part of the separation of CreateInviteLink into two components -> reimplemented the function here, but will remove it in a following diff.

Functionally should be the same.

Tried to move some tests into a gomock that's a bit nicer to use than what we have. Will follow up with more updates in this file if they are preferred.

Test Plan: Added some interface tests

Reviewers: vihang, michelle, vjain

Reviewed By: vihang

JIRA Issues: PC-1072

Differential Revision: https://phab.corp.pixielabs.ai/D9095

GitOrigin-RevId: 6d9f9c914a91d3aab0daf26458e591018a7ff925