PC-1072: Create identities in auth before creating in users table
Summary: As part of the series of diffs, we need to create identities in auth before creating the entry in our users table. This actually uses the methods implemented in the previous diff.

Test Plan:
Tested on skaffold with both create_admin_user and the invite flow (consumers of this api). Was able to use the recovery links in hydra to log back in.

To run it yourself, make sure you clear out the `user_settings, users, orgs, identities` tables in postgres and then deploy with
```skaffold dev -f skaffold/skaffold_cloud.yaml -p ory_auth --tail=false --no-prune```
( might need to push secrets as well. )

Reviewers: vihang, vjain

Reviewed By: vihang

JIRA Issues: PC-1072

Differential Revision: https://phab.corp.pixielabs.ai/D9105

GitOrigin-RevId: c8856407656f99d2bf4da33c03ed23cb09e71c22