PC-1073: Add InviteUser to auth service
Summary:
InviteUser is better suited for auth service instead of profile service. This is the first in a series of diffs to move this over. As part of the change, I'm also reducing the complexity of the API - removed the IdentityProvider field (can be inferred from auth service) and the MustCreateUser (not useful)

The following diffs will remove unnecessary deps from the profile service on idprovider and remove the implementation of InviteUser.

Test Plan: skffold dev with the ory_auth profile. Added tests

Reviewers: vihang

Reviewed By: vihang

JIRA Issues: PC-1073

Differential Revision: https://phab.corp.pixielabs.ai/D9117

GitOrigin-RevId: 0b73830bbf2f6f7355b54bae5a5d7cf9aab5fe93