PerfProfiler: Partially bring back stack IDs for efficiency
Summary:
When we removed stack trace IDs it saved memory, but it caused aggregations in PxL to become less efficient.

This partially restores stack trace IDs. There is no guarantee that an ID is unique across all time,
but it can be useful across shorter time periods.

Test Plan: Refactored and added a test.

Reviewers: #stirling, jps

Reviewed By: #stirling, jps

Subscribers: jps

Differential Revision: https://phab.corp.pixielabs.ai/D9118

GitOrigin-RevId: e03780477cf523d1c2b28613e366a689346cc899