Add csrf from cookie to request headers
Summary:
To protect against CSRF attacks, we are using the double submit cookie pattern.
Upon login, we will set the CSRF token in a cookie. The UI reads that cookie and attaches it to all requests.

Test Plan: yarn dev, check network tag and make sure x-csrf is set in the headers for our outgoing requests

Reviewers: zasgar, #third_party_approvers, vihang

Reviewed By: #third_party_approvers, vihang

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9119

GitOrigin-RevId: 18a2ed90b960db75eb2e03421e2f78740fd448f2