PC-1044: Add Unique constraint to auth_provider_id column in users table
Summary: auth_provider_id must be unique, setting that so in the database.

Test Plan: Tested on skaffold to make sure the query can be run. Also making sure that you can log in with a new user. Verified beforehand on prod/staging that you would log in with unique ids.

Reviewers: michelle, vihang, vjain

Reviewed By: vihang

JIRA Issues: PC-1044

Differential Revision: https://phab.corp.pixielabs.ai/D9125

GitOrigin-RevId: 2fc5f9d734c6695f8be660aa45be98f46f933a4e