[PP-2855] Fix memory leak in GRPCRouter.
Summary:
The GRPCRouter was implicitly creating QueryTrackers after the query had already been deleted by DeleteQuery. This leaks at least sizeof(QueryTracker) for every query that is cancelled or ends unexpectedly.
I changed the `query_node_map_` to store shared pointers to QueryTrackers instead of the object itself, so that QueryTrackers can no longer be implicitly created now or in the future. I also made it so that QueryTrackers are only created on calls to AddGRPCSourceNode or when a client calls TransferResultChunk with initiate_result_stream set. If the query is deleted but a client still calls TransferResultChunk, then the GRPCRouter will return an error to the client.

I also added a lock on each QueryTracker, since I think there was a race condition if multiple clients sent results for the same query at the same time.
EDIT: there was not actually a race in the code on master (because it happened to be holding the global lock whenever it made changes to a QueryTracker). But with my changes, the global lock is no longer held all over the place, so I needed to add a per query lock. To be sure, I added a TSAN test for the multi writer case.

Test Plan: Added a test for the leak scenario. Passes on my branch fails on main. Also tested to see if it leaks on skaffold and confirmed it doesn't.

Reviewers: zasgar, vihang, nserrino

Reviewed By: nserrino

JIRA Issues: PP-2855

Differential Revision: https://phab.corp.pixielabs.ai/D9137

GitOrigin-RevId: ab1469fd449b718319b7a9c46d7a63a8b6dcd8a2