[PC-???] Don't scroll the entire drawer, but rather its contents.
Summary:
When clicking a result row in the data drawer with large JSON (like in `px/http_data`), the entire data drawer was scrolling instead of its contents.

Without fix (note scrollbars in data drawer):
{F161861}

With fix (note scrollbars again):
{F161860}

Test Plan: Run `px/http_data`. Click on a row in the data drawer. The JSON representation in the sidebar that appears should now scroll independently of the table and the drawer's headers should stay put.

Reviewers: michelle, vihang

Reviewed By: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D9139

GitOrigin-RevId: 9b06477d1e923a98dde908c0e2706279cd5ecec2