[PC-1075] Restore default value for a variable if it gets emptied
Summary: When the user tries to clear a variable, if it has a default value, that value is restored. Previously it would try to run the script with that arg missing, which is an error.

Test Plan: Try changing `start_time` to an empty string in any script that uses it. It should be set to `-5m` instead. Other variables that lack a non-empty default stay blank (even if required) and will throw appropriate errors.

Reviewers: michelle, vihang

Reviewed By: vihang

JIRA Issues: PC-1075

Differential Revision: https://phab.corp.pixielabs.ai/D9147

GitOrigin-RevId: 099797da8343138bcdd14dfac9c8fee5da8c8467