Push permissions check for UpdateUser higher up
Summary:
we should check permissions in the API server. if the user is updating a user in the same org, then this should be allowed.
if the user is in the same org, but is updating their own isApproved, this should be blocked.
all other cases should be blocked

Test Plan: unit test

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9259

GitOrigin-RevId: 2fdb25eb3cb56d62fec02834a4cfac707e69c6ce