Embed view: send request to refresh token upon expiration
Summary:
We are currently running into an issue in the embed view where the view will stop working if it has been left open for over 30 mins.
We should send a message to the parent frame requesting that the token be refreshed. Our listener is still open, so we'll update it and things will start working again once they send it over.

Test Plan: yarn dev

Reviewers: vihang, zasgar

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9268

GitOrigin-RevId: cd66a87887f63a9c1e1ca65f4794d900117709c0