FPResolver: Store previously read mount infos for efficiency
Summary:
The ElfSymbolizer calls SetMountNamespace(pid) to switch between stack traces a lot.
In such cases, we don't want to keep re-reading the mount infos for a PID.

This new version optimizes by keeping previously read results and retrieving them on a future access.

Test Plan: This is a performance optimization with no functional changes, so rely on existing tests

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D9270

GitOrigin-RevId: fab61cf95e7cf13e0c2f1312eae475bd86f90ed7