PC-1090: Replace v2c subscribe with a Durable QueueSubscribe
Summary:
We want our consumer of v2c to be a Durable so that we don't have to re-ingest v2c messages that we've already processed (and thought to be the cause of PC-1090). We also need to make the subscriber a queue subscribe to handle the case where two vzmgr pods are running simultaneously during a release rollout (see comment on QueueSubscribe in indexer/md/md.go).

If you're comparing the two, it's important to note that we don't care about index updating for this QueueSubscribe so we don't use the Index name anywhere. Anything that's pushed on the Indexer stream can be reread by Indexer consumers By changing the durable names.

Test Plan: Tested on skaffold and updates still propagate as expected (update in cluster pod should show up in the autocomplete which uses data from the indexer).

Reviewers: vihang, michelle, vjain

Reviewed By: michelle

JIRA Issues: PC-1090

Differential Revision: https://phab.corp.pixielabs.ai/D9308

GitOrigin-RevId: 3ca414035fd43063426390e1adc29de84bc52c47