PC-1092 API Key should be associated to org, not user
Summary:
currently, when we create the claims for a user using an API key, we create the UserClaims for the user who created the API key. this can lead to problems if the user has been deleted. the API key should still be valid, since it is actually tied to the org.
allowed userClaims to be set as an "API user claim" using a boolean.

Test Plan: unit tests, deployed to staging and tested API key

Reviewers: zasgar, vihang

Reviewed By: zasgar

JIRA Issues: PC-1092

Differential Revision: https://phab.corp.pixielabs.ai/D9319

GitOrigin-RevId: 7871fb5cbeb7628498671b461ac717e2f6009a8f