Fix operator for dev cloud
Summary:
We were seeing issues where the operator could not connect to dev cloud. that is because it is using the incorrect cloudAddr.
we use this logic to set the cloudAddr for viziers already (see config service), but we were not updating the cloudAddr correctly for the operator itself.

technically, only the change in the operator is necessary. however, I made the additional change to the CLI since the operator needs to be heavily tested before its next release, so the CLI change will temporarily unblock for now

Test Plan: n/a

Reviewers: jamesbartlett

Reviewed By: jamesbartlett

Differential Revision: https://phab.corp.pixielabs.ai/D9341

GitOrigin-RevId: 098884b30bd3d0adb11ec51b54d56592cc7eecde