Add readiness probe to vzmgr
Summary:
MetadataReader is slow to startup and causes vzmgr to fail the
liveliness probe. Start serving requests without blocking on metadata reader
so that liveliness passes. Add a `readyz` handler that waits for metadatareader
so that we won't be added to the service loadbalancer until actually ready.

Test Plan: skaffold dev

Reviewers: michelle, philkuz

Reviewed By: michelle, philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D9365

GitOrigin-RevId: a8d93000c00f93faa10418979e4d109a0e718f60