Update operator release build to use kustomize labels
Summary:
we should be using the kustomized versions of our YAMLs to pass to OLM, since there are some common labels we want to add.
the missing labels were an issue when trying to run skaffold over a release version of the operator, since the skaffold version deploys the kustomized files with the labels.

Test Plan: built operator rc, then ran skaffold

Reviewers: vihang

Reviewed By: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D9394

GitOrigin-RevId: 162b78bb218e94da41508a3d7adcb50f852d6014